Explicit casts to ulong. Some compiler don't like pointers in %
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@14066 c90b9560-bf6c-de11-be94-00142212c4b1

+++ TSRM/TSRM.h
@@ -61,7 +61,7 @@ typedef int ts_rsrc_id;
 
 typedef void (*ts_allocate_ctor)(void *);
 
-#define THREAD_HASH_OF(thr,ts)  thr%ts
+#define THREAD_HASH_OF(thr,ts)  (unsigned long)thr%(unsigned long)ts
 
 #ifdef __cplusplus
 extern "C" {
